Correct default userModel
The user model is used by the auth component to know where in the
session to look for auth data. using $controller->modelClass as a
default means that the auth component looks in a different place for
each controller and therefore has the potential to invoke a redirect
loop where (example using an empty install):

you login - users controller, sessionKey in the auth component is set to
"Auth.User" derived from the userModel
you're redirected to the home.
you re-land at home - pages controller sessionKey in the auth component
is set to "Auth.Page" dervied from the userModel
you're redirected to login.

irgo: / -> /users/users/login -> / -> .. ad infinitum

After correcting this error, it is apparent that if you're logged in but
do not have access the auth component will redirect you to a "safe" url
which is assumed to be the home. If you have NOT granted access to /
this will again invoke a redirect loop this time / -> / as the auth
component tries to redirect you away from the page you don't have acces
for.