Rehashing fetched password
This fixes rehashing the password when the password is not passed in
the postData. The issue here is if a password was not passed, the user
is fetched from the datasource and then the hashed password is
rehashed. This results in rehashing a hash, which in turn unexpectedly
updates the password as a hash of a hash. Accompanied by test updating
single field and not modifying password.