unified style of hash-function
of the six calls of User->hash() some had `null`as second parameter and some had `sha1`. Both have the same effect, since the hash-method of Cake's Utility/Security-class defaults to `sha1` if the hashing-type is set to `null`. However, I think it is better to unify this, because if anything changes in the default-behaviour of Cake's hash-function this Plugin will break.