hash-function parameters set in one place
In order to reduce repetion, I removed the optional parameters from the six hash-calls. The parameters are now being set only once in the default-params definition.

This increases the readability of the code, especially when somebody overrides the User->hash() method to use a different hashing method, but doesn't want to change all six calls of the hash-method to remove the type and salt definitions there.