optional method for controlling public access
Please see the usage notes in the comments added to see how this is a new *optional* method allows users to control public or guest access from the permissions.php config file.  The benefit is that it allows all permissions to be controlled from a single place - including public permissions, as opposed to having some permissions in the individual controllers (eg. $this->Auth->allow('someAction')) and some permissions in the config file.