new method for controlling public access
This new method allows you to check public access and control all of your permissions from a single file. As opposed to the way it is now where you have to control your permissions from numerous different places when it comes to public and logged in user access.  eg. In each controller you would have to specify $this->Auth->allow('someAction'), but now you can just control it from the same place as all of your other permissions.