add parameters to UserShell::addSuperuser - issue #497
added same parameters to addSuperusers that already exists in addUser
method (-e, -p, -r, -u)
updated tests in UsersShellTest to test addSuperuser with and without
parameters.

addSuperuser and addUser are now more or less a repetition and should
be rewritten to keep it DRY. But I'll save that work for a rainy day.