Correct casing of users.po in translation docs
The users.po file which is used to overwrite the translations provided by the plugin should be lowercase. Seeing as though translation files are generally lowercase in CakePHP in addition to the fact that as soon as you deploy to a Linux environment this will bite you in the ass.